+++ tests/unit/base/test_commands_convert_onnx_onnx2trt.py
@@ -266,6 +266,7 @@ def test_run_execute_conversion_with_max_batch_size_search_when_both_max_batch_s
                 dataloader_trt_profile=TensorRTProfile(),
                 precision=TensorRTPrecision.FP16,
                 precision_mode=TensorRTPrecisionMode.HIERARCHY,
+                conversion_fallback=True,
                 custom_args={},
             )
 
@@ -274,6 +275,49 @@ def test_run_execute_conversion_with_max_batch_size_search_when_both_max_batch_s
             assert (
                 ConvertONNX2TRT._execute_conversion_with_max_batch_size_search.called
                 is True  # pytype: disable=attribute-error
+            )
+
+
+def test_run_execute_conversion_with_conversion_fallback_disabled(mocker):
+    with tempfile.TemporaryDirectory() as tmpdir:
+        tmpdir = pathlib.Path(tmpdir)
+        workspace = tmpdir / "navigator_workspace"
+
+        input_model_path = workspace / "onnx" / "model.onnx"
+        input_model_path.parent.mkdir(parents=True)
+        input_model_path.touch()
+
+        output_model_path = workspace / "trt-fp16" / "model.plan"
+        output_model_path.parent.mkdir(parents=True)
+
+        with (
+            mocker.patch.object(ConvertONNX2TRT, "_execute_single_conversion", return_value=3),
+            mocker.patch("model_navigator.utils.devices.get_available_gpus", return_value=[0]),
+        ):
+            result = ConvertONNX2TRT().run(
+                workspace=Workspace(workspace),
+                parent_path=input_model_path,
+                path=output_model_path,
+                input_metadata=TensorMetadata({
+                    "input__1": TensorSpec(name="input__1", shape=(-1,), dtype=np.dtype("float32"))
+                }),
+                output_metadata=TensorMetadata({
+                    "output__1": TensorSpec(name="output__1", shape=(-1,), dtype=np.dtype("float32"))
+                }),
+                batch_dim=0,
+                dataloader_max_batch_size=16,
+                device_max_batch_size=32,
+                dataloader_trt_profile=TensorRTProfile(),
+                precision=TensorRTPrecision.FP16,
+                precision_mode=TensorRTPrecisionMode.HIERARCHY,
+                conversion_fallback=False,
+                custom_args={},
+            )
+
+            assert result is not None
+            assert result.status == CommandStatus.OK
+            assert (
+                ConvertONNX2TRT._execute_single_conversion.called is True  # pytype: disable=attribute-error
             )